pyproject.toml
[project]
name = "password-reminder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "argon2-cffi>=25.1.0",
    "click>=8.3.1",
    "keyring>=25.7.0",
]

[project.scripts]
password-reminder = "main:main"

[dependency-groups]
dev = [
    "pyright>=1.1.408",
    "pytest>=9.0.2",
    "ruff>=0.15.8",
]
